Cleanup downloads directory after upgrade is run
The upgrader will delete the downloads directory after an upgrader is
run.

Previously, the GitHub upgrader would delete the files it
downloaded, and the NuGet upgrader would delete the extracted files (but
not the downloaded files). With this change, we will always delete the
download directory after upgrade is run.

If we decide to persist the download files for longer, that can be done
at a later time.

+++ GVFS/GVFS.UnitTests/Mock/FileSystem/MockFileSystem.cs
@@ -22,6 +22,8 @@ namespace GVFS.UnitTests.Mock.FileSystem
 
         public bool DeleteFileThrowsException { get; set; }
 
+        public Exception DeleteDirectoryExceptionToThrow { get; set; }
+
         public bool TryCreateOrUpdateDirectoryToAdminModifyPermissionsShouldSucceed { get; set; }
 
         /// <summary>
@@ -43,6 +45,11 @@ namespace GVFS.UnitTests.Mock.FileSystem
             if (!recursive)
             {
                 throw new NotImplementedException();
+            }
+
+            if (this.DeleteDirectoryExceptionToThrow != null)
+            {
+                throw this.DeleteDirectoryExceptionToThrow;
             }
 
             this.RootDirectory.DeleteDirectory(path);
